# 00-PROJECTS/HAIDA/haida-main/dev/docs/API/openapi.yaml
openapi: 3.0.3
info:
  title: HAIDA API
  description: |
    HAIDA - Quality Assurance Platform with AI Integration

    A comprehensive testing and auditing platform that combines traditional QA automation
    with AI-powered insights for enterprise software quality assurance.

    ## Key Features
    - Multi-tenant architecture with RBAC
    - Automated testing execution and monitoring
    - AI-powered chat and assistance (Copilot Studio)
    - Comprehensive documentation management
    - Real-time analytics and reporting
    - Enterprise SSO integration (Microsoft Entra ID)
  version: 2.0.0
  contact:
    name: HAIDA Development Team
    email: hola@stayarta.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://haida.vercel.app/api/v2
    description: Production server
  - url: http://localhost:8000/api/v2
    description: Development server
security:
  - bearerAuth: []
  - entraAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for authenticated requests
    entraAuth:
      type: oauth2
      flows:
        authorizationCode:
          authorizationUrl: https://login.microsoftonline.com/{tenant}/oauth2/v2.0/authorize
          tokenUrl: https://login.microsoftonline.com/{tenant}/oauth2/v2.0/token
          scopes:
            User.Read: Read user profile
            Files.Read.All: Read files
            Calendars.Read: Read calendar
            Mail.Read: Read mail
  schemas:
...
    UUID:
      type: string
      format: uuid
      example: "550e8400-e29b-41d4-a716-446655440000"
    Timestamp:
      type: string
      format: date-time
      example: "2025-12-16T10:30:00Z"
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
          minimum: 1
          default: 1
        limit:
          type: integer
          minimum: 1
          maximum: 100
          default: 20
        total:
          type: integer
          minimum: 0
        total_pages:
          type: integer
          minimum: 0
...
    Project:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/UUID'
        tenant_id:
          $ref: '#/components/schemas/UUID'
        name:
          type: string
          example: "E-commerce Platform"
        slug:
          type: string
          example: "ecommerce-platform"
        description:
          type: string
        base_url:
          type: string
          format: uri
          example: "https://shop.example.com"
        status:
          type: string
          enum: [active, inactive, archived]
          default: active
        type:
          type: string
          enum: [web, mobile, api, desktop]
          default: web
        settings:
          type: object
          properties:
            test_framework:
              type: string
              enum: [playwright, cypress, selenium]
              default: playwright
        created_by:
          $ref: '#/components/schemas/UUID'
        created_at:
          $ref: '#/components/schemas/Timestamp'
        updated_at:
          $ref: '#/components/schemas/Timestamp'
...
  parameters:
    tenantId:
      name: tenant_id
      in: header
      description: Tenant/workspace identifier (UUID or slug)
      required: false
      schema:
        type: string
    page:
      name: page
      in: query
      description: Page number for pagination
      schema:
        type: integer
        minimum: 1
        default: 1
    limit:
      name: limit
      in: query
      description: Number of items per page
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    search:
      name: search
      in: query
      description: Search query string
      schema:
        type: string
...
paths:
...
  /projects:
    get:
      summary: List projects
      description: Get all projects in the current tenant
      tags: [projects]
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/tenantId'
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/limit'
        - $ref: '#/components/parameters/search'
        - name: status
          in: query
          schema:
            type: string
            enum: [active, inactive, archived]
      responses:
        '200':
          description: Projects retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Project'
                  pagination:
                    $ref: '#/components/schemas/Pagination'

    post:
      summary: Create project
      description: Create a new project
      tags: [projects]
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/tenantId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, slug, base_url]
              properties:
                name:
                  type: string
                  minLength: 2
                  maxLength: 100
                slug:
                  type: string
                  pattern: '^[a-z0-9-]+$'
                description:
                  type: string
                base_url:
                  type: string
                  format: uri
                repository_url:
                  type: string
                  format: uri
                type:
                  type: string
                  enum: [web, mobile, api, desktop]
      responses:
        '201':
          description: Project created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
  /projects/{projectId}:
    get:
      summary: Get project details
      tags: [projects]
      security:
        - bearerAuth: []
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
      responses:
        '200':
          description: Project details retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'

    patch:
      summary: Update project
      tags: [projects]
      security:
        - bearerAuth: []
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                description:
                  type: string
                base_url:
                  type: string
                  format: uri
                status:
                  type: string
                  enum: [active, inactive, archived]
      responses:
        '200':
          description: Project updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'

  # Test Suites
...