# 00-PROJECTS/HAIDA/haida-production/main/haida/backend/app/api/v1/projects.py
"""
Project management API endpoints.

Endpoints:
- GET /api/v1/projects - List projects
- POST /api/v1/projects - Create project
- GET /api/v1/projects/{id} - Get project details
- PUT /api/v1/projects/{id} - Update project
- DELETE /api/v1/projects/{id} - Delete project (soft delete)
- GET /api/v1/projects/{id}/config - Get project configuration
- PUT /api/v1/projects/{id}/config - Update project configuration
"""

import uuid
from typing import Optional, List
from fastapi import APIRouter, Depends, HTTPException, Query
from sqlalchemy.orm import Session

from app.models.project import Project, ProjectConfig
from app.schemas.project import (
    ProjectCreateSchema,
    ProjectResponseSchema,
    ProjectListResponseSchema,
    ProjectUpdateSchema,
    ProjectConfigUpdateSchema,
    ProjectConfigGetResponseSchema,
)
from app.services.encryption import get_encryption_service

router = APIRouter(prefix="/api/v1/projects", tags=["projects"])


def get_db() -> Session:
    """Get database session. Replace with actual DB dependency."""
    # This is a placeholder. In real implementation, use SQLAlchemy session
    pass


def verify_owner(project: Project, user_id: str) -> None:
    """Verify that user is the project owner or admin."""
    if project.owner_id != user_id and user_id != "admin":
        raise HTTPException(status_code=403, detail="Not authorized to modify this project")


@router.get("", response_model=ProjectListResponseSchema)
def list_projects(
    skip: int = Query(0, ge=0),
    limit: int = Query(10, ge=1, le=100),
    is_active: Optional[bool] = None,
):
    """
    List all projects with optional filtering.

    Args:
        skip: Number of records to skip
        limit: Number of records to return
        is_active: Filter by active status
    """
    # This is a placeholder implementation
    # In real implementation, query database and return paginated results
    return ProjectListResponseSchema(
        items=[],
        total=0,
        skip=skip,
        limit=limit,
    )


@router.post("", response_model=ProjectResponseSchema)
def create_project(
    project_data: ProjectCreateSchema,
    current_user_id: str = "test-user",  # Replace with actual auth
):
    """
    Create a new project.

    Args:
        project_data: Project creation data
        current_user_id: Current user ID

    Returns:
        Created project
    """
    project_id = str(uuid.uuid4())

    # In real implementation:
    # project = Project(
    #     id=project_id,
    #     name=project_data.name,
    #     description=project_data.description,
    #     owner_id=current_user_id,
    #     is_active=project_data.is_active,
    #     config=project_data.config or {},
    # )
    # db.add(project)
    # db.commit()

    # Placeholder response
    return ProjectResponseSchema(
        id=project_id,
        name=project_data.name,
        description=project_data.description,
        owner_id=current_user_id,
        is_active=project_data.is_active,
        config=project_data.config or {},
        created_at=None,
        updated_at=None,
    )


@router.get("/{project_id}", response_model=ProjectResponseSchema)
def get_project(project_id: str):
    """
    Get project details by ID.

    Args:
        project_id: Project ID

    Returns:
        Project details
    """
    # In real implementation:
    # project = db.query(Project).filter(Project.id == project_id, Project.is_deleted == False).first()
    # if not project:
    #     raise HTTPException(status_code=404, detail="Project not found")
    # return project

    raise HTTPException(status_code=404, detail="Project not found")


@router.put("/{project_id}", response_model=ProjectResponseSchema)
def update_project(
    project_id: str,
    update_data: ProjectUpdateSchema,
    current_user_id: str = "test-user",
):
    """
    Update project details.

    Args:
        project_id: Project ID
        update_data: Update data
        current_user_id: Current user ID

    Returns:
        Updated project
    """
    # In real implementation:
    # project = db.query(Project).filter(Project.id == project_id).first()
    # if not project:
    #     raise HTTPException(status_code=404, detail="Project not found")
    #
    # verify_owner(project, current_user_id)
    #
    # if update_data.name:
    #     project.name = update_data.name
    # if update_data.description is not None:
    #     project.description = update_data.description
    # if update_data.is_active is not None:
    #     project.is_active = update_data.is_active
    # if update_data.config:
    #     project.config.update(update_data.config)
    #
    # db.commit()
    # return project

    raise HTTPException(status_code=404, detail="Project not found")


@router.delete("/{project_id}")
def delete_project(
    project_id: str,
    current_user_id: str = "test-user",
):
    """
    Delete project (soft delete).

    Args:
        project_id: Project ID
        current_user_id: Current user ID

    Returns:
        Deletion confirmation
    """
    # In real implementation:
    # project = db.query(Project).filter(Project.id == project_id).first()
    # if not project:
    #     raise HTTPException(status_code=404, detail="Project not found")
    #
    # verify_owner(project, current_user_id)
    # project.is_deleted = True
    # db.commit()

    return {"message": "Project deleted successfully", "project_id": project_id}


@router.get("/{project_id}/config", response_model=ProjectConfigGetResponseSchema)
def get_project_config(
    project_id: str,
    environment: str = Query("dev", description="Environment to get config for"),
):
    """
    Get project configuration for a specific environment.

    Args:
        project_id: Project ID
        environment: Environment name (dev, staging, prod)

    Returns:
        Configuration for the environment
    """
    # In real implementation:
    # project = db.query(Project).filter(Project.id == project_id, Project.is_deleted == False).first()
    # if not project:
    #     raise HTTPException(status_code=404, detail="Project not found")
    #
    # configs = db.query(ProjectConfig).filter(
    #     ProjectConfig.project_id == project_id,
    #     ProjectConfig.environment == environment
    # ).all()
    #
    # encryption_service = get_encryption_service()
    # config_dict = {}
    # for cfg in configs:
    #     if cfg.is_encrypted:
    #         config_dict[cfg.key] = encryption_service.decrypt(cfg.value)
    #     else:
    #         config_dict[cfg.key] = cfg.value
    #
    # return ProjectConfigGetResponseSchema(
    #     project_id=project_id,
    #     environment=environment,
    #     configs=config_dict,
    #     created_at=project.created_at,
    #     updated_at=project.updated_at,
    # )

    raise HTTPException(status_code=404, detail="Project not found")


@router.put("/{project_id}/config", response_model=ProjectConfigGetResponseSchema)
def update_project_config(
    project_id: str,
    config_data: ProjectConfigUpdateSchema,
    current_user_id: str = "test-user",
):
    """
    Update project configuration for a specific environment.

    Args:
        project_id: Project ID
        config_data: Configuration update data
        current_user_id: Current user ID

    Returns:
        Updated configuration
    """
    # In real implementation:
    # project = db.query(Project).filter(Project.id == project_id, Project.is_deleted == False).first()
    # if not project:
    #     raise HTTPException(status_code=404, detail="Project not found")
    #
    # verify_owner(project, current_user_id)
    #
    # encryption_service = get_encryption_service()
    #
    # for key, value in config_data.configs.items():
    #     # Check if config exists
    #     config = db.query(ProjectConfig).filter(
    #         ProjectConfig.project_id == project_id,
    #         ProjectConfig.environment == config_data.environment,
    #         ProjectConfig.key == key
    #     ).first()
    #
    #     # Determine if value should be encrypted
    #     is_sensitive = key.lower() in ["password", "token", "secret", "api_key"]
    #
    #     if config:
    #         if is_sensitive:
    #             config.value = encryption_service.encrypt(str(value))
    #             config.is_encrypted = True
    #         else:
    #             config.value = str(value)
    #             config.is_encrypted = False
    #         config.is_sensitive = is_sensitive
    #     else:
    #         config = ProjectConfig(
    #             id=str(uuid.uuid4()),
    #             project_id=project_id,
    #             environment=config_data.environment,
    #             key=key,
    #             value=encryption_service.encrypt(str(value)) if is_sensitive else str(value),
    #             is_encrypted=is_sensitive,
    #             is_sensitive=is_sensitive,
    #         )
    #         db.add(config)
    #
    #     db.commit()
    #
    # return get_project_config(project_id, config_data.environment)

    raise HTTPException(status_code=404, detail="Project not found")